pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["mian312 <deymilan312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.4.0"
python-dotenv = "^1.0.1"
vadersentiment = "^3.3.2"
speechrecognition = "^3.10.1"
pyaudio = "^0.2.14"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
setuptools = "^69.1.1"
gradio-client = "^0.11.0"
pymongo = "^4.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


